feat:relation-users-to-locals

+++ src/main/java/com/locals/locals_api/modules/locals/services/DeleteLocalsService.java
@@ -16,13 +16,16 @@ public class DeleteLocalsService {
     @Autowired
     private LocalsRepository localsRepository;
 
-    public void execute(UUID id) {
+    public void execute(UUID id, UUID userId) {
 
         if (!localsRepository.existsById(id)) {
             throw new RuntimeException("LocalsEntity with ID " + id + " does not exist.");
         }
         Optional<LocalsEntity> existingLocalsEntity = localsRepository.findById(id);
-
+        System.out.println(userId);
+        if(!(existingLocalsEntity.get().getUser().getId().equals(userId))){
+            throw new RuntimeException("user not unauthorized");
+        }
         if (existingLocalsEntity.get().getImageName() != null) {
                 // Deleta a imagem anterior
                 ImageUploadUtil.deleteImage(existingLocalsEntity.get().getImageName());
